# Server/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "StockQuotes",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/truemogician/StockQuote/0.1",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
...
  "paths" : {
...
    "/api/config" : {
      "get" : {
        "tags" : [ "Config" ],
        "description" : "Get the configuration of given token or default one if token not provided",
        "operationId" : "GetConfig",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Configuration returned successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Configuration"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid parameters or payload"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "Unknown error on server"
          }
        }
      },
      "put" : {
        "tags" : [ "Config" ],
        "description" : "Create or replace the configuration of a token",
        "operationId" : "CreateConfig",
        "requestBody" : {
          "description" : "Configuration data",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Configuration has been saved"
          },
          "400" : {
            "description" : "Invalid parameters or payload"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "Unknown error on server"
          }
        },
        "security" : [ {
          "token" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "Config" ],
        "description" : "Modify part of the configuration of a token",
        "operationId" : "ModifyConfig",
        "requestBody" : {
          "description" : "Part of the configuration that needs updating",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Configuration"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Configuration has been updated"
          },
          "400" : {
            "description" : "Invalid parameters or payload"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "Unknown error on server"
          }
        },
        "security" : [ {
          "token" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Configuration" : {
        "type" : "object",
        "properties" : {
          "pinnedStocks" : {
            "type" : "array",
            "items" : {
              "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
              "type" : "string"
            }
          },
          "refreshInterval" : {
            "$ref" : "#/components/schemas/Configuration_refreshInterval"
          }
        },
        "example" : {
          "refreshInterval" : {
            "single" : 6,
            "list" : 0
          },
          "pinnedStocks" : [ "pinnedStocks", "pinnedStocks" ]
        }
      },
...
      "body" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Configuration"
        }, {
          "required" : [ "pinnedStocks", "refreshInterval" ],
          "type" : "object"
        } ]
      },
      "Configuration_refreshInterval" : {
        "type" : "object",
        "properties" : {
          "list" : {
            "type" : "integer",
            "format" : "int32"
          },
          "single" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "example" : {
          "single" : 6,
          "list" : 0
        }
      }
    },
...
    "securitySchemes" : {
      "token" : {
        "type" : "apiKey",
        "name" : "token",
        "in" : "header"
      }
    }
  }
}